Add class 'popup' to inner div in noscript tag
We added the class 'popup' to the noscript tag to style the inner
message saying that JavaScript is disabled.  But we noticed that when
JavaScript is disabled the browser will not use the class on the
noscript tag.  This caused the text to be unstyled.  This commit fixes
that and adds the class 'popup' to a div inside the noscript tags,
instead of directly to the noscript tag.